Create shared remote settings client and port Nimbus to it.
