+++ Cargo.toml
@@ -23,6 +23,7 @@ members = [
     "components/support/rc_crypto/nss/nss_build_common",
     "components/support/rc_crypto/nss/nss_sys",
     "components/support/rc_crypto/nss/systest",
+    "components/support/rs_client",
     "components/support/rust-log-forwarder",
     "components/support/sql",
     "components/support/types",
@@ -91,6 +92,7 @@ default-members = [
     "components/support/rc_crypto/nss",
     "components/support/rc_crypto/nss/nss_build_common",
     "components/support/rc_crypto/nss/nss_sys",
+    "components/support/rs_client",
     "components/support/sql",
     "components/support/types",
     "components/support/viaduct-reqwest",
