+++ components/nimbus/src/tests/client/test_http_client.rs
@@ -2,14 +2,10 @@
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::client::{http_client::*, SettingsClient};
-use crate::{
-    error::NimbusError, Branch, BucketConfig, Experiment, FeatureConfig, RandomizationUnit,
-    RemoteSettingsConfig, SCHEMA_VERSION,
-};
+use crate::client::SettingsClient;
+use crate::{Branch, BucketConfig, Experiment, FeatureConfig, RandomizationUnit, SCHEMA_VERSION};
 use mockito::mock;
-use std::cell::Cell;
-use std::time::{Duration, Instant};
+use rs_client::{Client, ClientConfig};
 
 #[test]
 fn test_fetch_experiments_from_schema() {
@@ -24,9 +20,10 @@ fn test_fetch_experiments_from_schema() {
     .with_status(200)
     .with_header("content-type", "application/json")
     .create();
-    let config = RemoteSettingsConfig {
-        server_url: mockito::server_url(),
+    let config = ClientConfig {
+        server_url: Some(mockito::server_url()),
         collection_name: "messaging-experiments".to_string(),
+        bucket_name: None,
     };
     let http_client = Client::new(config).unwrap();
     let resp = http_client.fetch_experiments().unwrap();
@@ -77,85 +74,6 @@ fn test_fetch_experiments_from_schema() {
             ..Default::default()
         }
     )
-}
-
-#[test]
-fn test_backoff() {
-    viaduct_reqwest::use_reqwest_backend();
-    let m = mock(
-        "GET",
-        "/v1/buckets/main/collections/messaging-experiments/records",
-    )
-    .with_body(response_body())
-    .with_status(200)
-    .with_header("content-type", "application/json")
-    .with_header("Backoff", "60")
-    .create();
-    let config = RemoteSettingsConfig {
-        server_url: mockito::server_url(),
-        collection_name: "messaging-experiments".to_string(),
-    };
-    let http_client = Client::new(config).unwrap();
-    assert!(http_client.fetch_experiments().is_ok());
-    let second_request = http_client.fetch_experiments();
-    assert!(matches!(second_request, Err(NimbusError::BackoffError(_))));
-    m.expect(1).assert();
-}
-
-#[test]
-fn test_500_retry_after() {
-    viaduct_reqwest::use_reqwest_backend();
-    let m = mock(
-        "GET",
-        "/v1/buckets/main/collections/messaging-experiments/records",
-    )
-    .with_body("Boom!")
-    .with_status(500)
-    .with_header("Retry-After", "60")
-    .create();
-    let config = RemoteSettingsConfig {
-        server_url: mockito::server_url(),
-        collection_name: "messaging-experiments".to_string(),
-    };
-    let http_client = Client::new(config).unwrap();
-    assert!(http_client.fetch_experiments().is_err());
-    let second_request = http_client.fetch_experiments();
-    assert!(matches!(second_request, Err(NimbusError::BackoffError(_))));
-    m.expect(1).assert();
-}
-
-#[test]
-fn test_backoff_recovery() {
-    viaduct_reqwest::use_reqwest_backend();
-    let m = mock(
-        "GET",
-        "/v1/buckets/main/collections/messaging-experiments/records",
-    )
-    .with_body(response_body())
-    .with_status(200)
-    .with_header("content-type", "application/json")
-    .create();
-    let config = RemoteSettingsConfig {
-        server_url: mockito::server_url(),
-        collection_name: "messaging-experiments".to_string(),
-    };
-    let mut http_client = Client::new(config).unwrap();
-    // First, sanity check that manipulating the remote state does something.
-    http_client.remote_state.replace(RemoteState::Backoff {
-        observed_at: Instant::now(),
-        duration: Duration::from_secs(30),
-    });
-    assert!(matches!(
-        http_client.fetch_experiments(),
-        Err(NimbusError::BackoffError(_))
-    ));
-    // Then do the actual test.
-    http_client.remote_state = Cell::new(RemoteState::Backoff {
-        observed_at: Instant::now() - Duration::from_secs(31),
-        duration: Duration::from_secs(30),
-    });
-    assert!(http_client.fetch_experiments().is_ok());
-    m.expect(1).assert();
 }
 
 fn response_body() -> String {
