+++ components/support/rs_client/src/lib.rs
@@ -1,0 +1,10 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+* License, v. 2.0. If a copy of the MPL was not distributed with this
+* file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+mod http_client;
+pub use http_client::Client;
+mod error;
+pub use self::error::ClientError;
+mod config;
+pub use self::config::ClientConfig;
